Added state stack
This makes it possible for commands to return to the previous mode and
not always to normal mode.  For example, in Vim, if you're in visual
mode and you press `fa` to go to the next letter 'a', you should return
to visual and not go back to normal mode.