Fixed last character getting truncated from cmdlines
The last character on a cmdline would get replaced with a space when a
character was inserted at the beginning of the line.

Also, more stuff than necessary would get copied if the cursor wasn't at
the end of the line.  This could have probably been a source of
segfaults, but I didn't check to see if it actually was.